agent-prod-alpha/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "Radiance"
version = "1.0.0"
description = "Voice Agent to revolutionize healthcare in Pakistan"
requires-python = ">=3.9"

dependencies = [
    "livekit-agents[openai,turn-detector,silero,cartesia,deepgram,aws,azure,google,speechify,rime,resemble,lmnt,groq,neuphonic,hume,inworld,sarvam,spitch,assemblyai,speechmatics,anthropic,gladia,baseten,fal,elevenlabs]~=1.2",
    "livekit-plugins-noise-cancellation~=0.2",
    "python-dotenv",
    "httpx",
]

[dependency-groups]
dev = [
    "pytest",
    "pytest-asyncio",
    "ruff",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[tool.ruff]
line-length = 88
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "B", "A", "C4", "UP", "SIM", "RUF"]
ignore = ["E501"]  # Line too long (handled by formatter)

[tool.ruff.format]
quote-style = "double"
indent-style = "space"